chore(core/rust): Split FormattedText into separate mod
