+++ core/embed/rust/src/ui/component/text/mod.rs
@@ -1,0 +1,2 @@
+pub mod formatted;
+pub mod layout;
