+++ core/embed/rust/src/ui/model_t1/theme.rs
@@ -1,5 +1,5 @@
 use crate::ui::{
-    component::text::DefaultTextTheme,
+    component::text::layout::DefaultTextTheme,
     display::{Color, Font},
 };
